Rename kubernetes package to extractor and clean up
